requirements.txt
flask-cors
google-generativeai
langchain-google-genai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
pymupdf
pandas
httpx
asyncio
pytz
watchdog
python-dotenv
sentence-transformers
tqdm
rich
youtube-transcript-api
beautifulsoup4
selenium
pillow
flask
streamlit
pytesseract
scikit-learn
python-nmap
ephem
paho-mqtt
pyaudio
pvporcupine
ollama